requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
argon2-cffi==23.1.0
pydantic[email]==2.5.2
python-dotenv==1.0.0
uuid_v7==1.0.0
sqlmodel==0.0.29
psycopg[binary]==3.1.18
